requirements.txt
Bootstrap-Flask==2.4.0
email_validator==2.2.0
Flask==3.0.3
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fuzzywuzzy==0.18.0
Jinja2==3.1.3
marshmallow==3.21.3
PyJWT==2.8.0
python-Levenshtein==0.25.1
requests==2.32.3
SQLAlchemy==2.0.19
Werkzeug==3.0.1
WTForms==3.0.1